requirements.txt
ipykernel
openai==1.6.1
pandas==2.1.4
streamlit==1.31.1
python-dotenv==1.0.0
pocbot==0.3.0
chromadb==0.4.18
fastapi==0.109.2
uvicorn==0.27.1
watchdog
